+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/).
 
+## [3.0.0]
+
+### Added
+
+- esbuild integration (the one used by default).
+- rollup integration (optionally used).
+- webpack integration (optionally used).
+
 ## [2.4.1]
 
 ### Fixed